Add extensions for import to comply with ESM standard, test will be buggy otherwise
