+++ backend/src/game/game.plugin.ts
@@ -1,6 +1,6 @@
 import type { FastifyInstance } from "fastify";
-import { createGameService } from "./game.service";
-import gameRoutes from "./game.routes";
+import { createGameService } from "./game.service.ts";
+import gameRoutes from "./game.routes.ts";
 import fp from "fastify-plugin";
 
 const gamePlugin = async (fastify: FastifyInstance) => {
