+++ backend/src/tournament/tournament.plugin.ts
@@ -1,6 +1,6 @@
 import type { FastifyInstance } from "fastify";
-import { createTournamentService } from "./tournament.service";
-import tournamentRoutes from "./tournament.routes";
+import { createTournamentService } from "./tournament.service.ts";
+import tournamentRoutes from "./tournament.routes.ts";
 import fp from "fastify-plugin";
 
 const tournamentPlugin = async (fastify: FastifyInstance) => {
