+++ backend/src/user/__tests__/user.service.test.ts
@@ -1,7 +1,10 @@
 import { test } from "tap";
+import { createUserService } from "../user.service.ts";
 import { createUserSchema } from "../user.type.ts";
 
-test("createUserSchema - valid data passes", (t) => {
+test("createUserSchema - valid data passes", async (t) => {
+    // const app = await buildTestApp();
+
     const result = createUserSchema.safeParse({
         username: "darren",
         displayName: "Darren Dev",
